requirements.txt
anyio==4.12.1
cachetools==5.5.2
certifi==2026.2.25
charset-normalizer==3.4.6
google-auth==2.34.0
google-auth-oauthlib==1.3.0
gspread==6.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
oauthlib==3.3.1
pyaes==1.6.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
python-dotenv==1.0.1
python-telegram-bot==21.5
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
Telethon==1.36.0
urllib3==2.6.3
